requirements.txt
Flask==2.2.3
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.6
SQLAlchemy==2.0.10
python-dotenv==1.0.0
pydub==0.25.1